Maze Game
Created a simple maze game where the user navigates with the arrow
keys, through a randomly generated maze, only being able to see a 1
block radius around him. The goal is to find all of the tokens before
the monster(a randomly moving piece) finds you. Created from scratch
for a school project.